pyproject.toml
[project]
name = "fear-and-greed-algo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "moccasin>=0.4.2",
    "pytest-mock>=3.15.1",
    "requests>=2.32.5",
]
